pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "circuit-array-spec"
version = "0.1.0"
description = "Parser, validation and derivation helpers for the Circuit Array Specification"
readme = "README.md"
requires-python = ">=3.10"
authors = [{ name = "Generator" }]
dependencies = [
    "pydantic>=2.6,<3",
    "jsonschema>=4.0",
    "pytest-cov>=7.1.0",
    "Pillow>=10,<12",
]


[project.scripts]
circuit-array-debug = "arraylayout.debug.cli:main"
circuit-array-render = "arraylayout.render.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
